requirements.txt
python-geohash
shapely
geojson
jupyter
numpy
matplotlib